refactoring socket into it's own module
