+++ src/modules/index.ts
@@ -1,2 +1,5 @@
 export { default as Auth } from './auth';
 export { default as Booth } from './booth';
+
+// internal
+export { default as Socket } from './socket';
